move searchbar and sortby into main grid & algin searchbar with packages
